backend/requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# Machine Learning Core Libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
tensorflow>=2.13.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0

chromadb>=0.4.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openai>=1.0.0

# Hugging Face Ecosystem
accelerate>=0.20.0
diffusers>=0.18.0
sentence-transformers>=2.2.0
evaluate>=0.4.0

# Scientific Computing
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0
scikit-learn>=1.2.0

# Deep Learning Utilities
pytorch-lightning>=2.0.0
wandb>=0.15.0
tensorboard>=2.13.0
optuna>=3.0.0

# Computer Vision
opencv-python>=4.7.0
pillow>=9.0.0
albumentations>=1.3.0
timm>=0.9.0

# Natural Language Processing
spacy>=3.6.0
nltk>=3.8.0
textblob>=0.17.0

# Vector Databases and Search
chromadb>=0.4.0
faiss-cpu>=1.7.0
pinecone-client>=2.2.0

# Data Visualization
matplotlib>=3.6.0
seaborn>=0.11.0
plotly>=5.15.0
bokeh>=3.0.0

# API and External Services
openai>=1.0.0
anthropic>=0.3.0
requests>=2.28.0
httpx>=0.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0
# sqlite3 is built-in

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Jupyter and Notebooks
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# Model Optimization
onnx>=1.14.0
onnxruntime>=1.15.0